Prefix alarm names
This commit adds "haproxy_" as a prefix to alarm and alarm cluster
names. This greatly reduces the risk of name clashes.

We actually have a clash today, with two alarm clusters named
"rabbitmq_cluster", one from the haproxy formula and one from the
rabbitmq formula. Because of this clash rabbitmq_cluster is UNKNOWN in
Nagios. This commit fixes the problem.