Introduce port range feature
This is usefull in listen blocks for definition of multiple servers
that differs only by port number in port range block. This situation
can be result of multiple single-thread servers deployed in multi-core
environment to better utilize the available cores.

For example five contrail-api workers occupy ports 9100-9104.
This can be achieved by using port_range_length in the pillar,
port_range_length: 5 in this case.
For skipping first worker (worker_id 0), because it has other
responsibilities and to avoid overloading it by http requests
use the port_range_start_offset in the pillar,
port_range_start_offset: 1 in this case, it will only use ports
9101-9104 (skipping 9100).

PROD-18175

Change-Id: Ia1ac04660e0876f58c48a287c2d0c2adf9231a1c